Rename: class Command --> Commands::Trust
